+++ common_libs/nx_network/src/nx/network/upnp/upnp_device_searcher.h
@@ -144,7 +144,7 @@ private:
     mutable QnMutex m_mutex;
     quint64 m_timerID;
     nx::utils::AsyncOperationGuard m_handlerGuard;
-    std::map< QString, std::map< SearchHandler*, uint > > m_handlers;
+    std::map< QString, std::map< SearchHandler*, uintptr_t > > m_handlers;
     mutable QSet<QnInterfaceAndAddr> m_interfacesCache;
     //map<local interface ip, socket>
     std::map<QString, SocketReadCtx> m_socketList;
